Arrange Code
~ Change `chunk::destroy()` to `chunk::~chunk()`.
+ Add `Fonts/` to .gitignore.
~ Change `World::worldname` to `World::WorldName`.
~ Change `updateThreadRun` from `Definitions.cpp` to `GameView.cpp`
because it is only referenced from there.
~ Format the members in enum{} in `Definitions.h`.
~ Support UNIX Systems in `rnd()` because `RAND_MAX` will cause integer
overflow in UNIX systems.
~ Format method names in `GameView.cpp`.
- Remove `rebuiltChunksCount`, `updatedChunksCount`,
`unloadedChunksCount`.

+++ source/FunctionsKit.h
@@ -45,7 +45,7 @@ vector<string> split(string str, string pattern);
 
 inline double rnd()
 {
-    return (double)fastRand() / (RAND_MAX + 1);
+    return (double)fastRand() / ((double)RAND_MAX + 1);
 }
 
 inline int RoundInt(double d)
